pyproject.toml
[tool.poetry]
name = "nautobot-ecosystem-health"
version = "0.1.0"
description = "Tools to build Nautobot Ecosystem Health Dashboards"
authors = ["Cristian Sirbu"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.4"
ruamel-yaml = "^0.18.6"
rich = "^13.7.1"


[tool.poetry.group.dev.dependencies]
yamllint = "^1.35.1"
ruff = "^0.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
